src/requirements.txt
anyconfig==0.9.11
appnope==0.1.0
arrow==0.15.5
azure-common==1.1.25
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
backcall==0.1.0
binaryornot==0.4.4
boto3==1.13.1
botocore==1.16.1
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
cryptography==2.9.2
decorator==4.4.2
docutils==0.15.2
Flask==1.1.2
fsspec==0.6.3
google-api-core==1.17.0
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jinja2-time==0.2.0
jmespath==0.9.5
kedro==0.15.9
kedro-viz==3.2.0
MarkupSafe==1.1.1
mock==4.0.2
numexpr==2.7.1
numpy==1.18.3
oauthlib==3.1.0
opencv-python==4.2.0.34
pandas==0.25.3
pandas-gbq==0.13.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pip-tools==4.5.1
poyo==0.5.0
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyarrow==0.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydata-google-auth==1.1.0
Pygments==2.6.1
python-dateutil==2.8.0
python-json-logger==0.1.11
python-slugify==4.0.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3fs==0.4.0
s3transfer==0.3.3
semver==2.9.1
six==1.14.0
SQLAlchemy==1.3.16
tables==3.5.2
text-unidecode==1.3
toposort==1.5
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
xlrd==1.2.0
XlsxWriter==1.2.8
youtube-dl==2020.5.3
